feat: setup MongoDB for news and products data
