+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "@gsap/react": "^2.1.1",
     "gsap": "^3.12.5",
+    "mongodb": "^6.9.0",
     "next": "^14.2.4",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
