--- src/lib/data/newsData.js
@@ -1,101 +1,0 @@
-const newsData = [
-  {
-    newsId: "4KVMWjC",
-    tag: "活動",
-    publishedAt: "2024-05-30T07:00:00.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499573/ndnzr2uowv7z7l6l1vf6.png",
-    title: "Vlive Lab 即將舉辦 VTuber 線上交流聚會",
-    description: "促進合作交流的未來之旅",
-    content:
-      "在這個充滿可能性的虛擬世界裡，我們很榮幸邀請您參加一場前所未有的線上VTuber交流聚會。這個聚會的目的是為了促進VTuber之間的合作與交流，並為我們創造一個共同成長的平台。這次的聚會將會是一次讓我們一同展望未來的旅程。我們將共同探索如何在這個數位世界中建立緊密的合作關係，並找到彼此的共同點與可能的合作機會。我們特別邀請了一些業界知名的 VTuber 分享他們的成功之道，並提供寶貴的建議與指導。這是一個難得的機會，除了能與頂尖的 VTuber 面對面互動，還可以從他們的經驗中汲取靈感。同時，現場也將開辦一系列的工作坊和訓練課程，旨在幫助您提升自己的技能和專業知識。課程內容包含如何打造獨特的角色形象、如何有效地互動與吸引觀眾，以及如何在不同的平台上建立品牌。我們相信，透過這次聚會，將能夠建立起一個更加緊密的 VTuber 社群，並促進彼此之間的合作和支持。這個虛擬的世界給予了我們無限的可能性，而此次聚會將為您提供一個實現這些可能性的舞台。我們期待你的參與，讓我們一起在虛擬世界中創造美好的回憶！\r\n\r\n活動日期：2024年7月1日，星期五\r\n地點：zoom\r\n報名方式：請在我們的官方網站上報名。付款完成後，我們會以電子郵件發送活動的連結給您。",
-    url: "https://vlivelab.com/",
-  },
-  {
-    newsId: "jtuJOiV",
-    tag: "焦點",
-    publishedAt: "2024-05-21T02:22:00.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499574/uwlq4cxaukef1xuqhhcu.png",
-    title: "虛擬歌姬 MuiME 發布最新單曲",
-    description: "MuiME 最新單曲現正於各大串流平台熱播中",
-    content:
-      "睽違三年，MuiME 終於發布了最新單曲 《Monotropa-3》，有別以往一貫的強烈電子舞曲風格，這次與知名作曲家 Serenade S. 合作，改採用抒情搖滾的曲風，柔和的木吉他與輕快的鋼琴伴奏、配上 MuiME 溫柔內斂但又蘊藏爆發力的歌聲，讓人耳目一新。「這次對我來說是很特別的經驗。在過去三年內，我認為遇到了創作瓶頸，因此想要走出自己原本的舒適圈，探索不同的領域；所以選擇了以往沒有接觸過的曲風，希望能帶給聽眾全新的感受。」MuiME 在接受專訪時說道。\r\n\r\n甫推出一週，此曲目前在各大串流平台已累積 50 萬觀看次數。粉絲們盛讚 MuiME 的大膽嘗試，表示雖然不是熟悉的電子舞曲風格，但歌詞保有 MuiME 富有故事性與創意的特色，且其溫柔婉轉的歌聲與伴奏非常契合。音樂時報專欄作家 Megan Mitchell 評論：「MuiME 在這首曲子展現了她與眾不同的想法，抒情搖滾讓她的歌聲呈現出了一種全新的魅力。《Monotropa-3》散發出濃厚的情感，歌詞與 Serenade S. 譜出的動人旋律相得益彰，讓人不禁沉浸其中。這標誌著 MuiME 在音樂上的成長與多樣性，讓我們期待她未來更多的音樂作品。」對於粉絲們的熱情回應以及專業評論家的讚譽，MuiME 除了感謝並進一步表示，未來將繼續挑戰自我，探索音樂的無限可能性，並期待與更多優秀的音樂人合作，帶來更多令人驚艷的作品。《Monotropa-3》目前可於 YouTube Music、Spotify、Apple Music 上收聽，有興趣的讀者歡迎多多留意。",
-    url: "https://open.spotify.com/",
-  },
-  {
-    newsId: "DaHUQbh",
-    tag: "情報",
-    publishedAt: "2024-04-10T01:27:51.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499571/hmjxutwzajv18g8mvnpb.png",
-    title: "四重精靈合唱團宣布將於夏季登台演出",
-    description: "風靡全球的四重精靈合唱團即將來台",
-    content:
-      "以其空靈而動聽的阿卡貝拉聲音風靡全球的四重精靈合唱團，近日宣布將於今年 6 月開始他們的亞洲巡迴演唱會，其中第二站將會在台北流行音樂中心進行。這是他們自 2018 年成立以來首次在亞洲進行巡迴演唱，也是他們首次在台灣進行演出，無疑是對台灣樂迷的一大福音。\r\n\r\n四重精靈合唱團由女高音 Borneol、女中音 Lamiaceae、女低音 Olivinol 以及人聲打擊 Cobalt 四位成員組成，他們的音色各具特色，音域廣泛，歌聲清澈動人。他們的歌聲如同清澈的溪流，滋潤著人們的心靈，帶給聽眾一種深深的平靜和安慰。自從他們的首張專輯《Orthannel vi'less》問世以來，他們的歌曲中充滿了神秘的精靈語，每一首歌都讓粉絲們猜測其深層含義，這也成為他們的一大特色。他們的歌曲不僅深受樂迷喜愛，也在多個國家的音樂榜單上取得了優秀的成績，可謂是一鳴驚人。\r\n\r\n對於即將來到台灣進行演出，四重精靈合唱團的成員們表示非常期待。「因為我們來自寒冷的伊費塞拉奧凡諾（精靈國），所以我們都非常期待能夠來到溫暖的台灣，並且能夠和台灣的粉絲們交流互動。」他們說。當被問到來台灣有沒有什麼想做的事情時，他們異口同聲地用中文說「芒果冰！」，顯然他們對台灣的在地文化有著濃厚的興趣。他們也表示希望能夠在演唱會之餘，多了解台灣的文化和風俗。對於這次的演唱會，有興趣的讀者可以前往線上售票系統查看更多的資訊。讓我們一起期待四重精靈合唱團在台灣的精彩演出吧！",
-    url: "https://ticket.ibon.com.tw/Index/entertainment",
-  },
-  {
-    newsId: "kjWkVLx",
-    tag: "分析",
-    publishedAt: "2024-03-07T14:30:06.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499572/eijqr5aivvbdmsuz5npo.png",
-    title: "VTuber 產業：從創作到消費的全面解析",
-    description: "快速解析虛擬直播產業",
-    content:
-      "隨著科技的發展，虛擬實境已經逐漸融入我們的日常生活，而 VTuber（虛擬YouTuber）就是其中一個最具代表性的範例。VTuber 產業鏈包含了多個重要的環節，從創作、製作、營銷到最終的消費者，每一個環節都對整個產業鏈的運作有著關鍵性的影響。\r\n\r\n首先，創作是 VTuber 核心。這包含了角色設計、語音演譯等多個環節。在這個階段，創作者需要根據市場需求和自己的創新思維來設計出獨特且吸引人的 VTuber 角色和內容。接下來是製作環節，包括動畫製作、聲音錄製和後期製作等。在這個階段，專業的技術人員將創作者的想法轉化為實際的影像和聲音，並透過後期製作來完善和提升作品的整體質量。然後是營銷環節，這是 VTuber 產業鏈中非常重要的一環。營銷人員需要透過各種方式來推廣 VTuber，包括社交媒體、網路廣告、合作夥伴關係等。透過有效的營銷策略，VTuber能夠吸引更多的觀眾，從而提升自己的知名度和收入。最後，就是我們的消費者，他們是 VTuber 產業鏈的終點。消費者可以透過觀看VTuber的影片、參與直播互動、購買周邊商品等方式來消費 VTuber 的內容。消費者的反饋和需求也會影響到產業鏈的其他環節，進一步塑造 VTuber 的創作和行銷策略。\r\n\r\n總結來說，VTuber 產業鏈是一個複雜而緊密相連的系統，每一個環節都對整體的運作有著重要的影響。隨著虛擬實境技術的進步，我們可以期待 VTuber 產業會有更多的可能性和機會。",
-    url: "https://animecorner.me/",
-  },
-  {
-    newsId: "OoSEBeN",
-    tag: "情報",
-    publishedAt: "2024-02-15T13:56:10.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499569/wpx2wwemqs73uevnvcam.png",
-    title: "新一代 VTuber 角色肢體操作軟體 VTmotion Beta 即將問世",
-    description: "立即試用Beta版！",
-    content:
-      "一款全新的肢體操作軟體 - VTmotion，即將在市面上問世。這款軟體的目的是為了讓 VTuber 更方便地操作角色的肢體動作，提供更流暢且自然的動畫效果。這款專門為 VTuber 設計的軟體，使用最先進的動作捕捉技術，能精準地捕捉並轉換您的肢體動作，讓您的 VTuber 角色能更真實地呈現出各種動作。不論是微妙的手勢，還是大幅度的肢體動作，VTmotion 都能完美地捕捉並呈現。\r\n核心功能如下：1. 高精度動作捕捉：採用最先進的動作捕捉技術，能夠精準地捕捉每一個細微動作，讓角色動作更加自然流暢。\r\n2. 簡便操作介面：用戶友好的介面設計，讓操作變得更加簡單直觀，無需繁複的設置即可立刻開始使用。\r\n3. 多平台支持：兼容多種Vtuber平台，無論你使用的是哪個平台，這款軟體都能完美匹配。\r\n4. 動作實時反饋調整：即時反饋功能，讓你能夠即時查看並調整角色動作，確保每一個動作都完美無瑕。\r\n\r\n目前，VTmotion 已經開放 Beta 版本的試用。我們誠摯地邀請所有的 VTuber 參與 Beta 測試，提供您寶貴的反饋，幫助我們改進軟體，以滿足 VTuber 社群的需求。VTmotion 的 Beta 版本以可以在官方網站上免費下載，只需前往我們的官方網站，填寫簡單的申請表格，即可獲得試用資格。我們也設有專門的客服團隊，為您提供技術支援，解答您在使用過程中的任何問題。期待您的參與，一起體驗 VTmotion 的革新力量！",
-    url: "https://denchisoft.com/",
-  },
-  {
-    newsId: "YEcrUgJ",
-    tag: "情報",
-    publishedAt: "2024-01-06T06:01:00.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499564/wf8ayzpy6nsmikdau0gj.png",
-    title: "知名 Vsinger 團體「催化劑」開放招募新人！",
-    description: "下一個閃耀的新星也許就是你",
-    content:
-      "你有滿腔才華想尋找舞台發光嗎？以動人旋律、充滿活力的表演而聞名的 VTuber 團體「催化劑」今日宣布開始招募新一期中之人。此團體以單曲《電子交響夢》一戰成名，除了短時間內在 YouTube 上累積 30 萬訂閱，還曾受邀為國產劇情遊戲《庫哈諾辛山歷險記》演唱片頭曲，在 ACG 圈掀起話題。如今，這組備受期待的虛擬偶像團體正準備邁向新的里程碑。為了將獨特的音樂風格和精彩表演推向國際，催化劑決定擴大成員陣容，尋找同樣懷抱夢想的新血加入。\r\n\r\n「我們非常期待看到來自各方的人才加入。」催化劑計畫的創辦人 KK 說道，「催化劑的核心價值是突破界限，創造獨一無二的內容，所以我們渴望找到能歌善舞、勇敢展現自我的人。」而他們目前希望透過新人為團體注入新風貌，並進一步拓展國際市場。\r\n\r\n本次招募對象不限性別，只要年滿 18 歲以上，具有中華民國國籍，對唱歌和表演充滿熱情，並渴望在虛擬舞台上發光發熱，都歡迎報名參加。試鏡將分為多個階段，最終選拔將以線上直播方式進行，讓粉絲們也能參與這場盛事。詳細的報名截止日期、資格要求和試鏡流程，請參考催化劑官方網站及社群媒體專頁。別再猶豫，這個難得的機會，或許就是你邁向虛擬偶像之路的起點！",
-    url: "https://www.vsinger.tw/",
-  },
-  {
-    newsId: "xj44JKZ",
-    tag: "焦點",
-    publishedAt: "2023-12-25T07:00:00.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499575/j0vlzkxad38ujam5slmk.png",
-    title: "動漫展驚現 3 名神秘 coser！華麗造型引爆視覺盛宴",
-    description: "Coser 們出眾的顏值配合華麗的服裝驚豔群眾",
-    content:
-      "漫展大廳人潮湧動，五光十色的 Cosplay 服裝令人目不暇接。然而，在一片喧囂中，三個身影卻格外引人注目。他們是近期爆紅的 coser：九天、Valee、Zuya。扮演的角色來自一個鮮為人知的動漫－《未來偶像☆Project》(Mirai Idol☆Project)。\r\n\r\n走在最前面的是高大帥氣的九天，他cos的是作品中的男主角之一，羽城悠斗。有著紫色的短髮，頭髮兩側還有幾縷亮藍色的挑染，整個人看起來既神秘又充滿力量。他身穿花色外套，點綴著紅、白色花紋的華麗圖案，修身剪裁完美勾勒出他挺拔的身材。走在中間的是身材修長的 Valee，他 cos 的是另一位重要角色，煌星 (Kousei)，深邃的紫色短髮在燈光下閃爍著神秘的光澤。一身黑色襯托出他冷峻的氣質，更顯英姿颯爽。最後面是長相可愛的 Zuya，他 cos 的是風雅瑞希 (Fuga Mizuki)：一頭及肩的深粉色捲髮慵懶模樣，額前幾縷挑染成藍色的髮絲更顯俏皮；他穿著一件寬鬆的米白色皮夾克，圍著彩色絲巾。他們仔細研究角色的性格和服裝，以高水準的化妝和造型技巧還原了角色原型，彷彿作品中的角色真的穿越到現實之中，在漫展上吸引了眾多攝影師和參觀者的目光，引來陣陣驚喜尖叫和快門聲。他們擺出各種帥氣的姿勢，和粉絲們互動，分享他們對動漫的熱愛，為這次動漫展增添了一抹亮麗的色彩，也讓更多人感受到了 cosplay 的魅力。",
-    url: "https://www.ccpa.org.tw/tica/index.php?tpl=11",
-  },
-  {
-    newsId: "mfLr4w-",
-    tag: "活動",
-    publishedAt: "2023-12-13T08:48:07.000Z",
-    coverImg:
-      "https://res.cloudinary.com/dsme7klzf/image/upload/v1717499575/ffaimqdutqtk3xroaw1k.png",
-    title:
-      "InfiniteAI Inc.與 Galaxy Stars 聯手舉辦以社交舞會為主題的直播連動企劃",
-    description: "精彩直播與各項好禮等你來參加",
-    content:
-      "十位人氣爆表的 VTuber 嘉賓，分別是 InfiniteAI Inc. 的所屬藝人：琥珀韻、芽衣 (Mei)、翼夢，以及 Galaxy Stars 所屬藝人: 心愛 (Kokoa)、Riko、紗瓦Sawa、黒蝶 羽衣 (Kokuchou Hagoromo)、姫川 紫 (Himekawa Murasaki)、璃薰、Misa koo。VTuber 們將在本周末晚上八點齊聚一堂，以社交舞會為主題為觀眾們帶來一場絢麗的視聽盛宴。這場直播連動企劃將分為三個部分：在第一部分，每位 VTuber 會輪流上台獨舞表演，展示各具風格的舞技。第二部分是一個互動環節，觀眾們可以在線上與喜愛的 VTuber 共舞，通過虛擬角色與他們互動，共同創造出一場獨特的社交舞會體驗。這讓觀眾不僅僅是觀看，更能身臨其境地參與其中。最後部分將是一場特別的表演。十位  VTuber 聯手進行一場驚喜合作表演，展示他們的團隊合作精神和才華。這將是整個直播連動企劃的高潮。此場直播連動企劃可以通過 Youtube 平台觀看，除了在直播間與十位 VTuber 互動交流，參加到最後的觀眾還有機會贏得親筆簽名等豐富周邊禮物，這將是一場充滿活力、獨特而又令人難以忘懷的直播活動，相信將會為粉絲們帶來無限的歡樂和驚喜。",
-    url: "https://vchama.xyz/",
-  },
-];
-
-export default newsData;
